♻ refactor: 새로 구현한 파일 리더를 통해 주식 데이터 수집

+++ src/main/java/com/ant/juchumjuchum/scraper/koreastock/KoreaStockService.java
@@ -1,16 +1,17 @@
 package com.ant.juchumjuchum.scraper.koreastock;
 
 
+import static java.util.Collections.emptyList;
+
+import com.ant.juchumjuchum.scraper.koreastock.file.KosdaqParser;
+import com.ant.juchumjuchum.scraper.koreastock.file.KospiParser;
+import com.ant.juchumjuchum.scraper.koreastock.file.StockFileReader;
 import com.ant.juchumjuchum.stock.StockRepository;
 import com.ant.juchumjuchum.stock.domain.Stock;
 import com.ant.juchumjuchum.utils.FileUtil;
 import jakarta.annotation.PostConstruct;
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
@@ -23,19 +24,16 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @RequiredArgsConstructor
 public class KoreaStockService {
+    public static final String KOSPI_CODE_MST = "kospi_code.mst";
+    public static final String KOSDAQ_CODE_MST = "kosdaq_code.mst";
     private final KoreaStockDownloadService koreaStockDownloadService;
     private final StockRepository stockRepository;
 
     @Scheduled(cron = "0 0 0 * * MON-FRI")
     @Scheduled(cron = "0 0 9 * * MON-FRI")
     @PostConstruct
-    public void initStockInfo() {
-        initKosdaq();
-        initKospi();
-    }
-
-    private void initKosdaq() {
-        CompletableFuture.supplyAsync(() -> {
+    public CompletableFuture<Integer> initKosdaq() {
+        return CompletableFuture.supplyAsync(() -> {
             try {
                 List<Stock> kosdaq = downloadKosdaq();
                 stockRepository.saveAll(kosdaq);
@@ -47,8 +45,12 @@ public class KoreaStockService {
         });
     }
 
-    private void initKospi() {
-        CompletableFuture.supplyAsync(() -> {
+
+    @Scheduled(cron = "0 0 0 * * MON-FRI")
+    @Scheduled(cron = "0 0 9 * * MON-FRI")
+    @PostConstruct
+    public CompletableFuture<Integer> initKospi() {
+        return CompletableFuture.supplyAsync(() -> {
             try {
                 List<Stock> kospi = downloadKospi();
                 stockRepository.saveAll(kospi);
@@ -60,62 +62,36 @@ public class KoreaStockService {
         });
     }
 
-    List<Stock> downloadKosdaq() throws IOException {
-        Optional<File> kosdaq = koreaStockDownloadService.downloadStockInfo("kosdaq_code");
-        kosdaq.ifPresent(file -> {
-            FileUtil.unzipFile(file);
-            file.delete();
-        });
-        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(
-                new FileInputStream("kosdaq_code.mst"), "ms949"
-        ));
-        List<Stock> stocks = getKosdaqStocksFromMst(bufferedReader);
-        new File("./kosdaq_code.mst").delete();
-        return stocks;
-    }
-
-    public List<Stock> downloadKospi() throws IOException {
-        Optional<File> kospi = koreaStockDownloadService.downloadStockInfo("kospi_code");
-        kospi.ifPresent(file -> {
-            FileUtil.unzipFile(file);
-            file.delete();
-        });
-        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(
-                new FileInputStream("kospi_code.mst"), "ms949"
-        ));
-        List<Stock> stocks = getKospiStocksFromMst(bufferedReader);
-        new File("./kospi_code.mst").delete();
-        return stocks;
-    }
-
-    private List<Stock> getKosdaqStocksFromMst(BufferedReader bufferedReader) throws IOException {
-        String line;
-        List<Stock> stocks = new ArrayList<>();
-        while ((line = bufferedReader.readLine()) != null) {
-            Stock stock = Stock.builder()
-                    .id(line.substring(0, 9).trim())
-                    .name(line.substring(21, line.length() - 221).trim())
-                    .group(line.substring(line.length() - 221, line.length() - 219).trim())
-                    .build();
-            stocks.add(stock);
+    private List<Stock> downloadKosdaq() throws IOException {
+        Optional<File> downloadFile = koreaStockDownloadService.downloadStockInfo("kosdaq_code");
+        boolean unZipResult = unZipFile(downloadFile);
+        if (!unZipResult) {
+            return emptyList();
         }
-        bufferedReader.close();
+        StockFileReader fileReader = new StockFileReader(new KosdaqParser());
+        List<Stock> stocks = fileReader.readStockFile(KOSDAQ_CODE_MST);
+        new File(KOSDAQ_CODE_MST).delete();
         return stocks;
     }
 
-
-    private List<Stock> getKospiStocksFromMst(BufferedReader bufferedReader) throws IOException {
-        String line;
-        List<Stock> stocks = new ArrayList<>();
-        while ((line = bufferedReader.readLine()) != null) {
-            Stock stock = Stock.builder()
-                    .id(line.substring(0, 9).trim())
-                    .name(line.substring(21, line.length() - 227).trim())
-                    .group(line.substring(line.length() - 227, line.length() - 225).trim())
-                    .build();
-            stocks.add(stock);
+    private List<Stock> downloadKospi() throws IOException {
+        Optional<File> downloadFile = koreaStockDownloadService.downloadStockInfo("kospi_code");
+        boolean unZipResult = unZipFile(downloadFile);
+        if (!unZipResult) {
+            return emptyList();
         }
-        bufferedReader.close();
+        StockFileReader fileReader = new StockFileReader(new KospiParser());
+        List<Stock> stocks = fileReader.readStockFile(KOSPI_CODE_MST);
+        new File(KOSPI_CODE_MST).delete();
         return stocks;
+    }
+
+    private boolean unZipFile(Optional<File> downloadFile) {
+        if (downloadFile.isEmpty()) {
+            return false;
+        }
+        File file = downloadFile.get();
+        FileUtil.unzipFile(file);
+        return file.delete();
     }
 }
